+++ src/main/kotlin/chromia/VersionedRellParsers.kt
@@ -4,6 +4,8 @@ import net.postchain.rellide.jetbrains.language.parser.v0_16_1.RellLexer as Rell
 import net.postchain.rellide.jetbrains.language.parser.v0_16_1.RellParser as RellParser0161
 import net.postchain.rellide.jetbrains.language.parser.v0_16_2.RellLexer as RellLexer0162
 import net.postchain.rellide.jetbrains.language.parser.v0_16_2.RellParser as RellParser0162
+import net.postchain.rellide.jetbrains.language.parser.v0_16_3.RellLexer as RellLexer0163
+import net.postchain.rellide.jetbrains.language.parser.v0_16_3.RellParser as RellParser0163
 import org.antlr.v4.runtime.*
 
 /**
@@ -17,6 +19,7 @@ object VersionedRellParsers {
     private val parsers: Map<RellVersion, (String) -> List<SyntaxError>> = mapOf(
         RellVersion(0, 16, 1) to entry(::RellLexer0161, ::RellParser0161, RellParser0161::file),
         RellVersion(0, 16, 2) to entry(::RellLexer0162, ::RellParser0162, RellParser0162::file),
+        RellVersion(0, 16, 3) to entry(::RellLexer0163, ::RellParser0163, RellParser0163::file),
     )
 
     fun supports(version: RellVersion): Boolean = version in parsers